requirements.txt
opencv-python == 4.8.1.78
tqdm == 4.67.1
langchain == 0.3.12
langchain-google-genai == 2.0.9
pillow == 11.0.0
openai == 1.57.4
langchain-openai == 0.2.12
langchain-community == 0.3.12
requests == 2.32.3
langchain-ollama == 0.2.3
ollama == 0.4.7
chromadb
numpy == 1.26.4
albumentations == 2.0.3
modelscope == 1.22.3
python-dotenv == 1.0.1
pyaudio == 0.2.14
addict == 2.4.0
datasets == 3.2.0
simplejson == 3.19.3
SoundFile == 0.13.1
wave == 0.0.2
keyboard == 0.13.5
